+++ service-implementation/src/main/java/com/liaison/mailbox/service/core/processor/DirectorySweeper.java
@@ -73,6 +73,7 @@ import java.util.Map;
 import static com.liaison.mailbox.MailBoxConstants.BYTE_ARRAY_INITIAL_SIZE;
 import static com.liaison.mailbox.MailBoxConstants.CONFIGURATION_CONNECTION_TIMEOUT;
 import static com.liaison.mailbox.MailBoxConstants.CONFIGURATION_SERVICE_BROKER_ASYNC_URI;
+import static com.liaison.mailbox.MailBoxConstants.CONFIGURATION_SOCKET_TIMEOUT;
 
 /**
  * DirectorySweeper
@@ -279,6 +280,7 @@ public class DirectorySweeper extends AbstractProcessor implements MailBoxProces
         }
 
         int connectionTimeout = MailBoxUtil.getEnvironmentProperties().getInt(CONFIGURATION_CONNECTION_TIMEOUT);
+        int socketTimeout = MailBoxUtil.getEnvironmentProperties().getInt(CONFIGURATION_SOCKET_TIMEOUT);
 
         try {
 
@@ -306,6 +308,7 @@ public class DirectorySweeper extends AbstractProcessor implements MailBoxProces
                     HTTPRequest request = HTTPRequest.post(serviceBrokerUri)
                             .header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON)
                             .connectionTimeout(connectionTimeout)
+                            .socketTimeout(socketTimeout)
                             .inputData(JAXBUtility.marshalToJSON(workTicket))
                             .outputStream(responseStream);