+++ RELEASE_NOTES.md
@@ -2,6 +2,8 @@
 ---------------
 * GMB-946 - Analyze why there are more timer threads.
 * GMB-947 - Relay errors to LENS Metadata.
+* GMB-954 - Query back to service broker to determine the sender name when creating a process.
+* GGT-1329 removing errant period from provisioning system calls to echo
 
 4.13.0
 ---------------